Added login and auth with JWT, Added/updated bash/python test scripts, Auth tests seems to work fine, But I shall keep testing... Next should be Pagination

+++ src/main/java/nkemrocks/anyteam_v1/dto/session/request/Session_Create_RequestDTO.java
@@ -19,7 +19,7 @@ public record Session_Create_RequestDTO(
         String sessionName,
 
         @NotNull(message = ERROR_NULL_TTL)
-        Integer ttl,
+        Long ttl,
 
         @JsonDeserialize(as = LinkedHashSet.class)
         @NotNull(message = ERROR_NULL_REQUIREMENTS)
