+++ src/main/java/nkemrocks/anyteam_v1/dto/session/response/Session_Create_ResponseDTO.java
@@ -8,7 +8,7 @@ import java.util.UUID;
 public record Session_Create_ResponseDTO(
         UUID sessionId,
         String sessionName,
-        Integer ttl,
+        Long ttl,
         @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd, HH:mm:ss", timezone = "UTC")
         Instant dateCreated
 ) {
