+++ src/main/java/nkemrocks/anyteam_v1/projection/Session_Details_Projection.java
@@ -6,7 +6,7 @@ import java.util.UUID;
 public interface Session_Details_Projection {
     UUID getSessionId();
     String getSessionName();
-    Integer getTtl();
+    Long getTtl();
     Instant getDateCreated();
     Instant getDateUpdated();
